Update cast error Category and Product Class

+++ src/main/java/com/alessandrof/workingSpring/entities/Category.java
@@ -22,7 +22,7 @@ public class Category implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	
 	@JsonIgnore
@@ -33,17 +33,17 @@ public class Category implements Serializable {
 		
 	}
 
-	public Category(Integer id, String name) {
+	public Category(Long id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
